requirements.txt
Django==1.10.1
djangorestframework==3.4.7
markdown==2.6.6
django-filter==0.15.0
django-phonenumber-field==1.0.0
